# Conflicts:
#	dev/PresentationLayer/GUI/Components/StockGUI.java
#	dev/PresentationLayer/GUI/SupplierGUI/AllSupplierFrame.java
#	dev/PresentationLayer/GUI/SupplierGUI/SupplierFrame.java
#	dev/PresentationLayer/GUI/SupplierGUI/SupplierGUI.java
#	dev/ServiceLayer/Supplier_Stock/UserService.java
#	out/production/SuperLy/BusinessLayer/Stock/Inventory.class
#	out/production/SuperLy/BusinessLayer/Stock/Item.class
#	out/production/SuperLy/BusinessLayer/Stock/Tests/DamagedTest.class
#	out/production/SuperLy/BusinessLayer/Stock/Tests/InventoryTest.class
#	out/production/SuperLy/BusinessLayer/Stock/Tests/ItemTest.class
#	out/production/SuperLy/BusinessLayer/Stock/Tests/OrderTest.class
#	out/production/SuperLy/BusinessLayer/Supplier/OrderController.class
#	out/production/SuperLy/BusinessLayer/Supplier_Stock/Util_Supplier_Stock.class
#	out/production/SuperLy/ServiceLayer/Stock/CategoryService.class
#	out/production/SuperLy/ServiceLayer/Stock/DamagedService.class
#	out/production/SuperLy/ServiceLayer/Stock/InventoryService.class
#	out/production/SuperLy/ServiceLayer/Stock/ItemService.class
#	out/production/SuperLy/ServiceLayer/Supplier/SupplierService.class
#	out/production/SuperLy/ServiceLayer/Supplier_Stock/Response.class
#	out/production/SuperLy/ServiceLayer/Supplier_Stock/UserService.class
#	out/production/SuperLy/stock_supplier_db.db

+++ dev/BusinessLayer/Supplier/Suppliers/SupplierBusiness.java
@@ -110,7 +110,7 @@ public abstract class  SupplierBusiness {
         return null;
     }
 
-    public void addProduct(int productNum, String productName, String manufacturer, int price, int maxAmount, LocalDate expiryDate) throws Exception {
+    public void addProduct(int productNum, String productName, String manufacturer, float price, int maxAmount, LocalDate expiryDate) throws Exception {
         if (getProduct(productName,manufacturer) != null)
             throw new Exception("product already exists.");
         else
